move eatables
