+++ src/lib/data/filterables.ts
@@ -11,11 +11,11 @@ import {
 	cluesShared,
 	wintertodt
 } from '../collectionLog';
-import { Eatables } from '../eatables';
 import { gracefulItems } from '../skilling/skills/agility';
 import Crafting from '../skilling/skills/crafting';
 import Fletching from '../skilling/skills/fletching';
 import resolveItems from '../util/resolveItems';
+import { Eatables } from './eatables';
 
 const ores = resolveItems([
 	`Copper ore`,
